small fixes for repos, every POJO is well mapped now

+++ src/main/java/ua/roman/flats/repositories/name/NameDistrickRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.name;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.name.NameDistrict;
 
-public interface NameDistrickRepository extends CrudRepository<NameDistrict, Long> {
+public interface NameDistrickRepository extends CrudRepository<NameDistrict, Integer> {
 }
