+++ src/main/java/ua/roman/flats/repositories/type/TypeLandlordRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.type;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.type.TypeLandlord;
 
-public interface TypeLandlordRepository extends CrudRepository<TypeLandlord, Long> {
+public interface TypeLandlordRepository extends CrudRepository<TypeLandlord, Integer> {
 }
