+++ src/main/java/ua/roman/flats/repositories/type/TypeMarketRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.type;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.type.TypeMarket;
 
-public interface TypeMarketRepository extends CrudRepository<TypeMarket, Long> {
+public interface TypeMarketRepository extends CrudRepository<TypeMarket, Integer> {
 }
